media: wake outputs and benchmark worker topology
Post coalesced runtime wakeups after successful media publication while retaining the periodic timer for selection, pacing, health, and cleanup. Keep timer cadence metrics separate from posted media visits and expose wakeup/coalescing counters.\n\nAdd a deterministic four-worker owner-balanced topology matrix with exact 64-bit FNV-1a naming, local versus routed placement, per-program feed/preroll pressure, and route-failure causes. Correct benchmark and unit helpers to use the production hash width.\n\nVerification: make -C tests/unit test; make nginx; make incarnation; make api-graph; make graph-conflict; make bench-worker-topology; make bench-fanout-delay. The topology matrix kept input and program frames equal with zero route-failure causes; feed evictions appeared only at the bounded retained-window limit. The wakeup benchmark measured p50/p95/p99/p99.9 fanout delay at 1 ms, versus the prior 64/128 ms timer floor.

+++ tests/unit/test_owner.c
@@ -118,7 +118,7 @@ test_stability(void)
 {
     ngx_str_t   app = str("live");
     ngx_str_t   stream = str("news");
-    uint32_t    hash = ngx_media_owner_hash(&app, &stream);
+    uint64_t    hash = ngx_media_owner_hash(&app, &stream);
     ngx_uint_t  slot2, slot4;
 
     TEST_CASE("ownership is stable across worker counts");
